fix(graphics): :bug: fix `new_vf` not registering
Even if the shader compilation or program linking processes failed, the
`GLGfxHandle` for the program is still a valid handle, and therefore
must correspond to a valid `GLHandle` stored in the `ProgramContainer`.

A sensible fix is to insert/register the GLHandle for the program into
the `ProgramContainer` before initializing (compile shaders and link
program), but when this fails, it will leave a unlinked program floating
around, and the error is not properly handled.

This commit fixes that issue by the above approach, and therefore still
having some problems. Despite that, a shader compilation error is a
fatal error, and it should not be considered to much within the design
of the game.

Of course, the game developer can utilize the underlying API to properly
handle these issues, simply by introducing some sort of `mpsc` channels to
receive the error, and thus the error can be properly handled.