fix(scene): :bug: better texture error handling
Initially, when the texture loading failed, the sender will send an
`Err(...)` message to the `JoinToken`. Only when the result is polled,
the error is properly handled (in this case it caused a chain error,
crashing the program).

A slightly better way to handle this is instead, send an Execute (in
this case, for simplicity I send an Error) `GameUserEvent` to handle the
error (much like the Ok route). This fix follows this approach.